+++ app/Plugin/Log.php
@@ -57,7 +57,7 @@ class Log {
 		$sql = 'CREATE TABLE ' . $wpdb->prefix . "propstack_logs (
             `id` bigint(20) unsigned NOT NULL AUTO_INCREMENT,
             `time` datetime DEFAULT '1970-01-01 00:00:00' NOT NULL,
-            `log` text NOT NULL,
+            `log` longtext NOT NULL,
             `md5` text NOT NULL,
             `category` varchar(40) NOT NULL DEFAULT '',
             `state` varchar(40) NOT NULL DEFAULT '',